refactor(web-mobile): split admin_scoring_modal.jsx into 4 cohesive modules (mp-zac3)
Extract the 2130-line monolith into focused siblings to reduce merge-conflict
churn. The thin entry (admin_scoring_modal.jsx) preserves the window bridge and
re-exports the complete 29-symbol surface unchanged so every existing caller
keeps resolving without modification.

- admin_scoring_autosave.jsx — autosave/sync infra (AUTOSAVE_DEBOUNCE_MS,
  SyncStatusPill, useDebouncedRunningWrite)
- admin_scoring_team.jsx — team helpers + TeamScoreEditorModal
- admin_scoring_individual.jsx — ScoreEditorModal (individual editor)
- admin_scoring_modal.jsx (thin entry, ~70 lines) — window bridge + barrel

Also adds:
- web-mobile/check-imports.mjs — dependency-free static cross-module import
  checker that catches the ESM blind spot (esbuild/vitest don't fail on a
  missing named export; only native browser ESM does). Wired to js/check-imports
  target in Makefile and run as part of js/validate.
- admin_scoring_modal_barrel.test.jsx — vitest barrel-completeness test that
  asserts all 29 named exports are present and window.ScoreEditorModal is set.
